Add stats info footer in Partners Stats page
- Create `StatsInfoFooter` component
