+++ frontend/src/components/partners/partnersStats.js
@@ -4,6 +4,7 @@ import shortNumber from 'short-number';
 import messages from './messages';
 import { StatsCard } from '../statsCard';
 import { MappingIcon, EditIcon, RoadIcon, HomeIcon } from '../svgIcons';
+import StatsInfoFooter from '../statsInfoFooter';
 
 const iconClass = 'h-50 w-50';
 const iconStyle = { height: '45px' };
@@ -43,31 +44,34 @@ export const StatsColumn = ({ label, value, icon }: Object) => {
 
 export const StatsSection = ({ partner }) => {
   return (
-    <div className="flex justify-between items-center" style={{ gap: '1.6rem' }}>
-      <StatsCard
-        icon={<MappingIcon className={iconClass} style={iconStyle} />}
-        description={<FormattedMessage {...messages.contributors} />}
-        value={partner ? partner.users : '-'}
-        className={'w-25-l w-50-m w-100 mv1'}
-      />
-      <StatsCard
-        icon={<EditIcon className={iconClass} style={iconStyle} />}
-        description={<FormattedMessage {...messages.editsStats} />}
-        value={partner ? partner.edits : '-'}
-        className={'w-25-l w-50-m w-100 mv1'}
-      />
-      <StatsCard
-        icon={<HomeIcon className={iconClass} style={iconStyle} />}
-        description={<FormattedMessage {...messages.buildingsStats} />}
-        value={partner ? partner.buildings : '-'}
-        className={'w-25-l w-50-m w-100 mv1'}
-      />
-      <StatsCard
-        icon={<RoadIcon className={iconClass} style={iconStyle} />}
-        description={<FormattedMessage {...messages.roadsStats} />}
-        value={partner ? partner.roads : '-'}
-        className={'w-25-l w-50-m w-100 mv1'}
-      />
+    <div className="mb2">
+      <div className="flex justify-between items-center" style={{ gap: '1.6rem' }}>
+        <StatsCard
+          icon={<MappingIcon className={iconClass} style={iconStyle} />}
+          description={<FormattedMessage {...messages.contributors} />}
+          value={partner ? partner.users : '-'}
+          className={'w-25-l w-50-m w-100 mv1'}
+        />
+        <StatsCard
+          icon={<EditIcon className={iconClass} style={iconStyle} />}
+          description={<FormattedMessage {...messages.editsStats} />}
+          value={partner ? partner.edits : '-'}
+          className={'w-25-l w-50-m w-100 mv1'}
+        />
+        <StatsCard
+          icon={<HomeIcon className={iconClass} style={iconStyle} />}
+          description={<FormattedMessage {...messages.buildingsStats} />}
+          value={partner ? partner.buildings : '-'}
+          className={'w-25-l w-50-m w-100 mv1'}
+        />
+        <StatsCard
+          icon={<RoadIcon className={iconClass} style={iconStyle} />}
+          description={<FormattedMessage {...messages.roadsStats} />}
+          value={partner ? partner.roads : '-'}
+          className={'w-25-l w-50-m w-100 mv1'}
+        />
+      </div>
+      <StatsInfoFooter />
     </div>
   );
 };
