+++ frontend/src/components/userDetail/elementsMapped.js
@@ -1,5 +1,5 @@
 import humanizeDuration from 'humanize-duration';
-import { useIntl, FormattedMessage } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
 import messages from './messages';
 import {
@@ -12,8 +12,7 @@ import {
   ValidatedIcon,
 } from '../svgIcons';
 import { StatsCard, DetailedStatsCard } from '../statsCard';
-import { useOsmStatsMetadataQuery } from '../../api/stats';
-import { dateOptions } from '../statsTimestamp';
+import StatsInfoFooter from '../statsInfoFooter';
 
 export const TaskStats = ({ userStats, username }) => {
   const {
@@ -114,7 +113,6 @@ export const shortEnglishHumanizer = humanizeDuration.humanizer({
 });
 
 export const ElementsMapped = ({ userStats, osmStats }) => {
-  const intl = useIntl();
   const duration = shortEnglishHumanizer(userStats.timeSpentMapping * 1000, {
     round: true,
     delimiter: ' ',
@@ -124,8 +122,6 @@ export const ElementsMapped = ({ userStats, osmStats }) => {
 
   const iconClass = 'h-50 w-50';
   const iconStyle = { height: '45px' };
-
-  const { data: osmStatsMetadata } = useOsmStatsMetadataQuery();
 
   return (
     <div>
@@ -177,22 +173,8 @@ export const ElementsMapped = ({ userStats, osmStats }) => {
           unitLess={osmStats?.waterway?.modified?.unit_less}
         />
       </div>
-      <div className="cf w-100 relative tr pt3">
-        <span className="ma0 f7 fw4 blue-grey mb1 i">
-          These statistics come from{' '}
-          <a
-            className="blue-grey fw7"
-            href="https://stats.now.ohsome.org/about"
-            target="_blank"
-            rel="noreferrer"
-          >
-            ohsomeNow Stats
-          </a>{' '}
-          and were last updated at{' '}
-          <strong>{intl.formatDate(osmStatsMetadata?.max_timestamp, dateOptions)}</strong> (
-          {intl.timeZone}).
-        </span>
-      </div>
+
+      <StatsInfoFooter />
     </div>
   );
 };
